refactor: ランキング機能をカスタムフックに分離してコードの可読性を向上
- useRankingData: ランキングデータの取得と状態管理を担当
- useRankingNavigation: 検索とページネーション機能をまとめて管理
- App.tsxが100行以上削減され、よりシンプルで保守しやすいコードに改善
- ロジックの関心の分離により、テストやデバッグが容易になった
